requirements.txt
# Note: recent Python and C++ versions require fasttext git https://github.com/facebookresearch/fastText#building-fasttext-for-python
fasttext @ git+https://github.com/facebookresearch/fastText@6c2204ba66776b700095ff73e3e599a908ffd9c3
transformers==4.56.1
minicons==0.3.18
datasets==2.18.0
evaluate==0.4.1
numpy==1.26.4
pandas==2.2.0
spacy==3.7.4
stanza==1.6.1
spacy-udpipe==1.0.0
spacy_stanza==1.0.4
spacy_conll==4.0.1
conllu==6.0.0
scipy==1.14.1
lexicalrichness==0.5.1
nltk==3.8.1
scikit-learn==1.5
pyphen==0.17.0
textstat==0.7.4
joblib==1.4.2
dill==0.3.8
seaborn==0.13.2
matplotlib==3.10.0
shap==0.46.0
statsmodels==0.14.4
torch==2.6.0
bitsandbytes==0.45.5
accelerate==1.6.0
outlines==0.2.3
sentencepiece==0.2.0
openai==1.77.0
python-dotenv==1.1.0
huggingface-hub==0.34.4
pytube==15.0.0
google-api-python-client==2.170.0
youtube-transcript-api==1.2.2
clam==3.2.11
surprisal==0.1.7
hf_xet==1.1.9
hf_transfer==0.1.9